requirements.txt
bme280==0.6
Flask==2.2.2
python_dateutil==2.8.2
RPi.bme280==0.2.4
smbus2==0.4.2
SQLAlchemy==1.4.41
